Start fixing things up to build with a proper cross-compiler.

+++ AK/StringBuilder.cpp
@@ -8,7 +8,7 @@ namespace AK {
 inline void StringBuilder::will_append(size_t size)
 {
     if ((m_length + size) > m_buffer.size())
-        m_buffer.grow(max(16u, m_buffer.size() * 2 + size));
+        m_buffer.grow(max((size_t)16, m_buffer.size() * 2 + size));
 }
 
 StringBuilder::StringBuilder(size_t initial_capacity)
